+++ R/marginaleffects.R
@@ -116,6 +116,7 @@
 #' arguments.
 #'
 #' @template model_specific_arguments
+#' @template bayesian
 #'
 #' @return A `data.frame` with one row per observation (per term/group) and several columns:
 #' * `rowid`: row number of the `newdata` data frame