+++ NEWS.md
@@ -1,7 +1,9 @@
-# marginaleffects 0.8.1.9000
+# marginaleffects 0.8.1.9001
 
 New features:
 
+* `hypothesis` can be used to set the null used in calculating Z and p. For example, is the conditional risk ratio different from 1?
+  - `comparisons(mod, transform_pre = "ratio", hypothesis = 1)`
 * Bayesian posterior distribution summaries can be customized using global options:
   - `options("marginaleffects_posterior_interval" = "eti")` 
   - `options("marginaleffects_posterior_interval" = "hdi")`